#65 - Moving PredicateUtil out to the public util package
