+++ core/src/main/java/com/adobe/aem/commons/assetshare/search/impl/SearchImpl.java
@@ -19,14 +19,13 @@
 
 package com.adobe.aem.commons.assetshare.search.impl;
 
-import com.adobe.aem.commons.assetshare.components.predicates.impl.PredicateUtil;
 import com.adobe.aem.commons.assetshare.search.Constants;
 import com.adobe.aem.commons.assetshare.search.Search;
 import com.adobe.aem.commons.assetshare.search.UnsafeSearchException;
 import com.adobe.aem.commons.assetshare.search.providers.SearchProvider;
 import com.adobe.aem.commons.assetshare.search.results.Results;
-import com.adobe.aem.commons.assetshare.search.results.impl.results.AbstractResultsImpl;
 import com.adobe.aem.commons.assetshare.search.results.impl.results.EmptyResultsImpl;
+import com.adobe.aem.commons.assetshare.util.PredicateUtil;
 import com.day.cq.wcm.api.Page;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
@@ -41,7 +40,6 @@ import org.slf4j.LoggerFactory;
 
 import javax.annotation.PostConstruct;
 import javax.jcr.RepositoryException;
-import java.util.Collections;
 import java.util.List;
 
 @Model(
