+++ core/src/main/java/com/adobe/aem/commons/assetshare/util/PredicateUtil.java
@@ -17,7 +17,7 @@
  *
  */
 
-package com.adobe.aem.commons.assetshare.components.predicates.impl;
+package com.adobe.aem.commons.assetshare.util;
 
 import com.adobe.cq.wcm.core.components.models.form.OptionItem;
 import org.apache.commons.lang3.ArrayUtils;
@@ -26,9 +26,18 @@ import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.request.RequestParameter;
 import org.apache.sling.api.resource.ValueMap;
 
-public class PredicateUtil {
+/**
+ * Utility class that helps with common patterns found in Predicate implementations.
+ */
+public final class PredicateUtil {
     private PredicateUtil() { }
 
+    /**
+     *
+     * @param request the request object.
+     * @param parameterName The request query parameter name.
+     * @return a String representation of the query parameter {@param parameterName}. If no request parameter can be found with that name, the empty string is returned.
+     */
     public static String getParamFromQueryParams(final SlingHttpServletRequest request, final String parameterName) {
         final RequestParameter requestParameter = request.getRequestParameter(parameterName);
 
@@ -39,11 +48,27 @@ public class PredicateUtil {
         }
     }
 
-    public static boolean isOptionInInitialValues(String groupId, OptionItem optionItem, ValueMap initialValues) {
-        return isOptionInInitialValues(groupId, optionItem.getValue(), initialValues);
+    /**
+     * Determines if the {@optionItem}'s value exists as a value in the {@initialValues} map.
+     *
+     * @param optionItem the option item.
+     * @param initialValues the initial values.
+     * @return true if {@param optionItem}'s value is in {@initialValues}.
+     */
+    public static boolean isOptionInInitialValues(OptionItem optionItem, ValueMap initialValues) {
+        return isOptionInInitialValues(optionItem.getValue(), initialValues);
     }
 
-    public static boolean isOptionInInitialValues(String groupId, String value, ValueMap initialValues) {
+    /**
+     * Determines if the {@param value} exists as a value in the initialValues map.
+     *
+     * Only String and String[] intialValues values are supported.
+     *
+     * @param value the value to check.
+     * @param initialValues the initial values.
+     * @return true if {@param value} is in {@initialValues}.
+     */
+    public static boolean isOptionInInitialValues(String value, ValueMap initialValues) {
         boolean found = false;
         for (final String key : initialValues.keySet()) {
             if (initialValues.get(key) instanceof String) {
